Styling and consistency with list view
