+++ packages/dataviews/src/dataviews-layouts/timeline/index.tsx
@@ -8,368 +8,25 @@ import clsx from 'clsx';
  */
 import { useInstanceId, usePrevious } from '@wordpress/compose';
 import {
-	__experimentalHStack as HStack,
 	__experimentalVStack as VStack,
-	Button,
-	privateApis as componentsPrivateApis,
 	Spinner,
-	VisuallyHidden,
 	Composite,
 } from '@wordpress/components';
-import {
-	useCallback,
-	useEffect,
-	useMemo,
-	useRef,
-	useState,
-	useContext,
-} from '@wordpress/element';
+import { useCallback, useEffect, useState } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
-import { moreVertical } from '@wordpress/icons';
-import { useRegistry } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
-import { unlock } from '../../lock-unlock';
-import {
-	ActionsMenuGroup,
-	ActionModal,
-} from '../../components/dataviews-item-actions';
-import DataViewsContext from '../../components/dataviews-context';
-import type {
-	Action,
-	NormalizedField,
-	ViewTimeline as ViewTimelineType,
-	ViewTimelineProps,
-	ActionModal as ActionModalType,
-} from '../../types';
+import type { ViewTimelineProps } from '../../types';
 import getDataByGroup from '../utils/get-data-by-group';
-
-interface ListViewItemProps< Item > {
-	view: ViewTimelineType;
-	actions: Action< Item >[];
-	idPrefix: string;
-	isSelected: boolean;
-	item: Item;
-	titleField?: NormalizedField< Item >;
-	mediaField?: NormalizedField< Item >;
-	descriptionField?: NormalizedField< Item >;
-	onSelect: ( item: Item ) => void;
-	otherFields: NormalizedField< Item >[];
-	onDropdownTriggerKeyDown: React.KeyboardEventHandler< HTMLButtonElement >;
-	posinset?: number;
-}
-
-const { Menu } = unlock( componentsPrivateApis );
-
-function generateItemWrapperCompositeId( idPrefix: string ) {
-	return `${ idPrefix }-item-wrapper`;
-}
-function generatePrimaryActionCompositeId(
-	idPrefix: string,
-	primaryActionId: string
-) {
-	return `${ idPrefix }-primary-action-${ primaryActionId }`;
-}
-function generateDropdownTriggerCompositeId( idPrefix: string ) {
-	return `${ idPrefix }-dropdown`;
-}
-
-function PrimaryActionGridCell< Item >( {
-	idPrefix,
-	primaryAction,
-	item,
-}: {
-	idPrefix: string;
-	primaryAction: Action< Item >;
-	item: Item;
-} ) {
-	const registry = useRegistry();
-	const [ isModalOpen, setIsModalOpen ] = useState( false );
-
-	const compositeItemId = generatePrimaryActionCompositeId(
-		idPrefix,
-		primaryAction.id
-	);
-
-	const label =
-		typeof primaryAction.label === 'string'
-			? primaryAction.label
-			: primaryAction.label( [ item ] );
-
-	return 'RenderModal' in primaryAction ? (
-		<div role="gridcell" key={ primaryAction.id }>
-			<Composite.Item
-				id={ compositeItemId }
-				render={
-					<Button
-						disabled={ !! primaryAction.disabled }
-						accessibleWhenDisabled
-						size="small"
-						onClick={ () => setIsModalOpen( true ) }
-						variant="link"
-					>
-						{ label }
-					</Button>
-				}
-			>
-				{ isModalOpen && (
-					<ActionModal< Item >
-						action={ primaryAction }
-						items={ [ item ] }
-						closeModal={ () => setIsModalOpen( false ) }
-					/>
-				) }
-			</Composite.Item>
-		</div>
-	) : (
-		<div role="gridcell" key={ primaryAction.id }>
-			<Composite.Item
-				id={ compositeItemId }
-				render={
-					<Button
-						disabled={ !! primaryAction.disabled }
-						accessibleWhenDisabled
-						size="small"
-						onClick={ () => {
-							primaryAction.callback( [ item ], { registry } );
-						} }
-						variant="link"
-					>
-						{ label }
-					</Button>
-				}
-			/>
-		</div>
-	);
-}
-
-function ListItem< Item >( {
-	view,
-	actions,
-	idPrefix,
-	isSelected,
-	item,
-	titleField,
-	mediaField,
-	descriptionField,
-	onSelect,
-	otherFields,
-	onDropdownTriggerKeyDown,
-	posinset,
-}: ListViewItemProps< Item > ) {
-	const {
-		showTitle = true,
-		showMedia = true,
-		showDescription = true,
-		infiniteScrollEnabled,
-	} = view;
-	const itemRef = useRef< HTMLDivElement >( null );
-	const labelId = `${ idPrefix }-label`;
-	const descriptionId = `${ idPrefix }-description`;
-
-	const registry = useRegistry();
-	const [ isHovered, setIsHovered ] = useState( false );
-	const [ activeModalAction, setActiveModalAction ] = useState(
-		null as ActionModalType< Item > | null
-	);
-	const handleHover: React.MouseEventHandler = ( { type } ) => {
-		const isHover = type === 'mouseenter';
-		setIsHovered( isHover );
-	};
-
-	const { paginationInfo } = useContext( DataViewsContext );
-	useEffect( () => {
-		if ( isSelected ) {
-			itemRef.current?.scrollIntoView( {
-				behavior: 'auto',
-				block: 'nearest',
-				inline: 'nearest',
-			} );
-		}
-	}, [ isSelected ] );
-
-	const { primaryAction, eligibleActions } = useMemo( () => {
-		// If an action is eligible for all items, doesn't need
-		// to provide the `isEligible` function.
-		const _eligibleActions = actions.filter(
-			( action ) => ! action.isEligible || action.isEligible( item )
-		);
-		const _primaryActions = _eligibleActions.filter(
-			( action ) => action.isPrimary
-		);
-		return {
-			primaryAction: _primaryActions[ 0 ],
-			eligibleActions: _eligibleActions,
-		};
-	}, [ actions, item ] );
-
-	const hasOnlyOnePrimaryAction = primaryAction && actions.length === 1;
-
-	const renderedMediaField =
-		showMedia && mediaField?.render ? (
-			<div className="dataviews-view-timeline__media-wrapper">
-				<mediaField.render
-					item={ item }
-					field={ mediaField }
-					config={ { sizes: '52px' } }
-				/>
-			</div>
-		) : null;
-
-	const renderedTitleField =
-		showTitle && titleField?.render ? (
-			<titleField.render item={ item } field={ titleField } />
-		) : null;
-
-	const usedActions = eligibleActions?.length > 0 && (
-		<HStack spacing={ 3 } className="dataviews-view-timeline__item-actions">
-			{ primaryAction && (
-				<PrimaryActionGridCell
-					idPrefix={ idPrefix }
-					primaryAction={ primaryAction }
-					item={ item }
-				/>
-			) }
-			{ ! hasOnlyOnePrimaryAction && (
-				<div role="gridcell">
-					<Menu placement="bottom-end">
-						<Menu.TriggerButton
-							render={
-								<Composite.Item
-									id={ generateDropdownTriggerCompositeId(
-										idPrefix
-									) }
-									render={
-										<Button
-											size="small"
-											icon={ moreVertical }
-											label={ __( 'Actions' ) }
-											accessibleWhenDisabled
-											disabled={ ! actions.length }
-											onKeyDown={
-												onDropdownTriggerKeyDown
-											}
-										/>
-									}
-								/>
-							}
-						/>
-						<Menu.Popover>
-							<ActionsMenuGroup
-								actions={ eligibleActions }
-								item={ item }
-								registry={ registry }
-								setActiveModalAction={ setActiveModalAction }
-							/>
-						</Menu.Popover>
-					</Menu>
-					{ !! activeModalAction && (
-						<ActionModal
-							action={ activeModalAction }
-							items={ [ item ] }
-							closeModal={ () => setActiveModalAction( null ) }
-						/>
-					) }
-				</div>
-			) }
-		</HStack>
-	);
-
-	return (
-		<Composite.Row
-			ref={ itemRef }
-			render={
-				/* aria-posinset breaks Composite.Row if passed to it directly. */
-				<div
-					aria-posinset={ posinset }
-					aria-setsize={
-						infiniteScrollEnabled
-							? paginationInfo.totalItems
-							: undefined
-					}
-				/>
-			}
-			role={ infiniteScrollEnabled ? 'article' : 'row' }
-			className={ clsx( {
-				'is-selected': isSelected,
-				'is-hovered': isHovered,
-			} ) }
-			onMouseEnter={ handleHover }
-			onMouseLeave={ handleHover }
-		>
-			<HStack
-				className="dataviews-view-timeline__item-wrapper"
-				spacing={ 0 }
-			>
-				<div role="gridcell">
-					<Composite.Item
-						id={ generateItemWrapperCompositeId( idPrefix ) }
-						aria-pressed={ isSelected }
-						aria-labelledby={ labelId }
-						aria-describedby={ descriptionId }
-						className="dataviews-view-timeline__item"
-						onClick={ () => onSelect( item ) }
-					/>
-				</div>
-				<HStack spacing={ 3 } justify="start" alignment="flex-start">
-					{ renderedMediaField }
-					<VStack
-						spacing={ 1 }
-						className="dataviews-view-timeline__field-wrapper"
-					>
-						<HStack spacing={ 0 }>
-							<div
-								className="dataviews-title-field"
-								id={ labelId }
-							>
-								{ renderedTitleField }
-							</div>
-							{ usedActions }
-						</HStack>
-						{ showDescription && descriptionField?.render && (
-							<div className="dataviews-view-timeline__field">
-								<descriptionField.render
-									item={ item }
-									field={ descriptionField }
-								/>
-							</div>
-						) }
-						<div
-							className="dataviews-view-timeline__fields"
-							id={ descriptionId }
-						>
-							{ otherFields.map( ( field ) => (
-								<div
-									key={ field.id }
-									className="dataviews-view-timeline__field"
-								>
-									<VisuallyHidden
-										as="span"
-										className="dataviews-view-timeline__field-label"
-									>
-										{ field.label }
-									</VisuallyHidden>
-									<span className="dataviews-view-timeline__field-value">
-										<field.render
-											item={ item }
-											field={ field }
-										/>
-									</span>
-								</div>
-							) ) }
-						</div>
-					</VStack>
-				</HStack>
-			</HStack>
-		</Composite.Row>
-	);
-}
-
-function isDefined< T >( item: T | undefined ): item is T {
-	return !! item;
-}
+import TimelineItem from './timeline-item';
+import {
+	generateItemWrapperCompositeId,
+	isDefined,
+	generateDropdownTriggerCompositeId,
+	getFormattedDate,
+} from './utils';
 
 export default function ViewTimeline< Item >(
 	props: ViewTimelineProps< Item >
@@ -387,6 +44,9 @@ export default function ViewTimeline< Item >(
 		empty,
 	} = props;
 	const baseId = useInstanceId( ViewTimeline, 'view-timeline' );
+
+	const { eventField = undefined, dateFormat = undefined } =
+		view?.layout ?? {};
 
 	const selectedItem = data?.findLast( ( item ) =>
 		selection.includes( getItemId( item ) )
@@ -552,10 +212,13 @@ export default function ViewTimeline< Item >(
 				>
 					{ Array.from( dataByGroup.entries() ).map(
 						( [ groupName, groupItems ] ) => (
-							<VStack key={ groupName } spacing={ 2 }>
+							<VStack key={ groupName } spacing={ 0 }>
 								<h3 className="dataviews-view-timeline__group-header">
-									{ view.layout?.eventField === groupField.id
-										? groupName
+									{ eventField === groupField.id
+										? getFormattedDate(
+												groupName,
+												dateFormat
+										  )
 										: sprintf(
 												// translators: 1: The label of the field e.g. "Date". 2: The value of the field, e.g.: "May 2022".
 												__( '%1$s: %2$s' ),
@@ -563,36 +226,30 @@ export default function ViewTimeline< Item >(
 												groupName
 										  ) }
 								</h3>
-								<div className="dataviews-view-timeline__group-items">
-									{ groupItems.map( ( item ) => {
-										const id =
-											generateCompositeItemIdPrefix(
-												item
-											);
-										return (
-											<ListItem
-												key={ id }
-												view={ view }
-												idPrefix={ id }
-												actions={ actions }
-												item={ item }
-												isSelected={
-													item === selectedItem
-												}
-												onSelect={ onSelect }
-												mediaField={ mediaField }
-												titleField={ titleField }
-												descriptionField={
-													descriptionField
-												}
-												otherFields={ otherFields }
-												onDropdownTriggerKeyDown={
-													onDropdownTriggerKeyDown
-												}
-											/>
-										);
-									} ) }
-								</div>
+								{ groupItems.map( ( item ) => {
+									const id =
+										generateCompositeItemIdPrefix( item );
+									return (
+										<TimelineItem
+											key={ id }
+											view={ view }
+											idPrefix={ id }
+											actions={ actions }
+											item={ item }
+											isSelected={ item === selectedItem }
+											onSelect={ onSelect }
+											mediaField={ mediaField }
+											titleField={ titleField }
+											descriptionField={
+												descriptionField
+											}
+											otherFields={ otherFields }
+											onDropdownTriggerKeyDown={
+												onDropdownTriggerKeyDown
+											}
+										/>
+									);
+								} ) }
 							</VStack>
 						)
 					) }
@@ -615,7 +272,7 @@ export default function ViewTimeline< Item >(
 				{ data.map( ( item, index ) => {
 					const id = generateCompositeItemIdPrefix( item );
 					return (
-						<ListItem
+						<TimelineItem
 							key={ id }
 							view={ view }
 							idPrefix={ id }
